Modul 3.5_poprawione

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -1,15 +1,16 @@
 package com.kodilla.basic_assertion;
 
 public class Calculator {
-    public int add(int a, int b) {
+
+    public double add(double a, double b) {
         return a + b;
     }
 
-    public int sub(int a, int b) {
+    public double sub(double a, double b) {
         return a - b;
     }
 
-    public int pow(int a, int b) {
-        return ((int) Math.pow(a, b)) - 1; //"-1" odjęte żeby metoda dawała zły wynik
+    public double pow(double a, double b) {
+        return (Math.pow(a, b)) - 1; //"-1" odjęte żeby metoda dawała zły wynik
     }
 }
